Parsing: get grammar from incomplete HG
In the two pass algorithm, sometimes the first pass may fail to
generate a hypergraph that spans the entire input sentence. It
used to be that we bailed out, but that doesn't give us any
information about what partial parses may exist.

Here, we still generate a second pass grammar (ala Dyer) by
walking hypergraphs rooted at each complete item in the first pass
chart. In the second pass, that allows us to generate possible
synchronous trees that cover subspans of the two sentences.

Note that this may not cover the entire second sentence, so we
need extra reporting to show what the "best" subtree was in that
case. This is subject to change.